+++ src/fides/api/ops/models/messaging.py
@@ -56,10 +56,10 @@ def get_schema_for_secrets(
     """
     try:
         schema = {
-            MessagingServiceType.MAILCHIMP_TRANSACTIONAL: MessagingServiceSecretsMailchimpTransactional,
-            MessagingServiceType.MAILGUN: MessagingServiceSecretsMailgun,
-            MessagingServiceType.TWILIO_TEXT: MessagingServiceSecretsTwilioSMS,
-            MessagingServiceType.TWILIO_EMAIL: MessagingServiceSecretsTwilioEmail,
+            MessagingServiceType.mailgun: MessagingServiceSecretsMailgun,
+            MessagingServiceType.twilio_text: MessagingServiceSecretsTwilioSMS,
+            MessagingServiceType.twilio_email: MessagingServiceSecretsTwilioEmail,
+            MessagingServiceType.mailchimp_transactional: MessagingServiceSecretsMailchimpTransactional,
         }[service_type]
     except KeyError:
         raise ValueError(
@@ -127,7 +127,7 @@ class MessagingConfig(Base):
         """
         Retrieve the messaging config of the given type
         """
-        return db.query(cls).filter_by(service_type=service_type).first()
+        return db.query(cls).filter_by(service_type=service_type.value).first()
 
     def set_secrets(
         self,